patch: errorHandling refactor & coverage support
