+++ src/controllers/stops.controller.ts
@@ -1,3 +1,4 @@
+import { createError, ErrorTypes } from "../handlers/errorGenerator.ts";
 import { getStops } from "../services/stops/stops.service.ts";
 
 // deno-lint-ignore no-explicit-any
@@ -7,8 +8,9 @@ export async function getStopByLinea(ctx: any, next:any) {
         const stops = await getStops(linea)
         ctx.response.body = stops;  
     } catch (error) {
-        ctx.response.status = error.status || 500
-        ctx.response.body = error.body || error.message
+        const generic = createError(ErrorTypes.InternalError, error)
+        ctx.response.status = error.status || generic.status 
+        ctx.response.body = error.body || generic.body
         next()
     }
