+++ src/interfaces/Error.ts
@@ -1,0 +1,9 @@
+export interface ServiceError {
+    status: number;
+    body: Body;
+  }
+  
+interface Body {
+message: string;
+apiResponse: Record<string | number | symbol, never>;
+}
